docs: added missing info about pre-requirements about using customer-managed keys
I added missing info about pre-requirements about using customer-managed keys. I was looking for this statement 2 days.
It should be added to as a pre-requisite in every place where there is information about restore of Azure SQL managed Instance that restoring  it anywhere else than in the same Azure tenant is NOT possible with service-managed keys. Only customer-managed keys are supported in this scenario. Everywhere I see scenarios with using customer-managed keys but no information about what to do when using service-managed keys